RAC: use new Input component
(cherry picked from commit d05e8c4f141bc384ce1b048a955e8cb78467edb6)
