+++ packages/components/src/NumberField/NumberField.tsx
@@ -5,7 +5,7 @@ import { Group, NumberField } from 'react-aria-components';
 import { WidthProp, cn, useClassNames } from '@marigold/system';
 
 import { FieldBase, FieldBaseProps } from '../FieldBase/_FieldBase';
-import { Input } from '../Input/_Input';
+import { Input } from '../Input/Input';
 import { StepButton } from './StepButton';
 
 // Props
